Update scala, scala native, model, shared, fix prometheus test

+++ core/src/main/scalanative/sttp/client3/internal/CryptoMd5.scala
@@ -3,6 +3,7 @@ package sttp.client3.internal
 import scala.scalanative.unsafe._
 import scala.scalanative.libc.string.strlen
 import scala.scalanative.runtime.ByteArray
+import scala.scalanative.unsigned._
 
 object CryptoMd5 {
   @link("crypto")
@@ -13,7 +14,7 @@ object CryptoMd5 {
 
   def digest(input: Array[Byte]): Array[Byte] = {
     val result = ByteArray.alloc(16)
-    C.MD5(input.asInstanceOf[ByteArray].at(0), input.length, result.at(0))
+    C.MD5(input.asInstanceOf[ByteArray].at(0), input.length.toUInt, result.at(0))
     result.asInstanceOf[Array[Byte]]
   }
 }
